+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-03 15:36:34
-// Commit: refactor(core): decouple repository interface from concrete core service
+// Updated at: 2026-08-03 22:48:24
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '3304';
-  final String lastMessage = 'refactor(core): decouple repository interface from concrete core service';
-  final DateTime timestamp = DateTime.parse('2026-08-03T15:36:34+05:00');
+  final String commitRef = '3335';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-08-03T22:48:24+05:00');
 }